Polishing - PR Comments
Move utilities that don't depend on the container to s-k-test.

Document the test jar.
