+++ src/reference/asciidoc/testing.adoc
@@ -1,0 +1,131 @@
+[[testing]]
+=== Testing Applications
+
+==== Introduction
+
+The `spring-kafka-test` jar contains some useful utilities to assist with testing your applications.
+
+==== JUnit
+
+`o.s.kafka.test.utils.KafkaUtils` provides some static methods to set up producer and consumer properties:
+
+[source, java]
+----
+/**
+ * Set up test properties for an {@code <Integer, String>} consumer.
+ * @param group the group id.
+ * @param autoCommit the auto commit.
+ * @param embeddedKafka a {@link KafkaEmbedded} instance.
+ * @return the properties.
+ */
+public static Map<String, Object> consumerProps(String group, String autoCommit, KafkaEmbedded embeddedKafka) { ... }
+
+/**
+ * Set up test properties for an {@code <Integer, String>} producer.
+ * @param embeddedKafka a {@link KafkaEmbedded} instance.
+ * @return the properties.
+ */
+public static Map<String, Object> senderProps(KafkaEmbedded embeddedKafka) { ... }
+----
+
+A JUnit `@Rule` is provided that creates an embedded kafka server.
+
+[source, java]
+----
+/**
+ * Create embedded Kafka brokers.
+ * @param count the number of brokers.
+ * @param controlledShutdown passed into TestUtils.createBrokerConfig.
+ * @param topics the topics to create (2 partitions per).
+ */
+public KafkaEmbedded(int count, boolean controlledShutdown, String... topics) { ... }
+
+/**
+ *
+ * Create embedded Kafka brokers.
+ * @param count the number of brokers.
+ * @param controlledShutdown passed into TestUtils.createBrokerConfig.
+ * @param partitions partitions per topic.
+ * @param topics the topics to create.
+ */
+public KafkaEmbedded(int count, boolean controlledShutdown, int partitions, String... topics) { ... }
+----
+
+==== Hamcrest Matchers
+
+The `o.s.kafka.test.hamcrest.KafkaMatchers` provides the following matchers:
+
+[source, java]
+----
+/**
+ * @param key the key
+ * @return a Matcher that matches the key in a consumer record.
+ */
+public static <K> Matcher<ConsumerRecord<K, ?>> hasKey(K key) { ... }
+
+/**
+ * @param value the value.
+ * @return a Matcher that matches the value in a consumer record.
+ */
+public static <V> Matcher<ConsumerRecord<?, V>> hasValue(V value) { ... }
+
+/**
+ * @param partition the partition.
+ * @return a Matcher that matches the partition in a consumer record.
+ */
+public static Matcher<ConsumerRecord<?, ?>> hasPartition(int partition) { ... }
+----
+
+==== Example
+
+Putting it all together:
+
+[source, java]
+----
+public class KafkaTemplateTests {
+
+	private static final String TEMPLATE_TOPIC = "templateTopic";
+
+	@ClassRule
+	public static KafkaEmbedded embeddedKafka = new KafkaEmbedded(1, true, TEMPLATE_TOPIC);
+
+	@Test
+	public void testTemplate() throws Exception {
+		Map<String, Object> consumerProps = KafkaTestUtils.consumerProps("testT", "false", embeddedKafka);
+		DefaultKafkaConsumerFactory<Integer, String> cf = new DefaultKafkaConsumerFactory<Integer, String>(
+				consumerProps);
+		KafkaMessageListenerContainer<Integer, String> container = new KafkaMessageListenerContainer<>(cf,
+				TEMPLATE_TOPIC);
+		final BlockingQueue<ConsumerRecord<Integer, String>> records = new LinkedBlockingQueue<>();
+		container.setMessageListener(new MessageListener<Integer, String>() {
+
+			@Override
+			public void onMessage(ConsumerRecord<Integer, String> record) {
+				System.out.println(record);
+				records.add(record);
+			}
+
+		});
+		container.setBeanName("templateTests");
+		container.start();
+		ContainerTestUtils.waitForAssignment(container, embeddedKafka.getPartitionsPerTopic());
+		Map<String, Object> senderProps = KafkaTestUtils.senderProps(embeddedKafka);
+		ProducerFactory<Integer, String> pf = new DefaultKafkaProducerFactory<Integer, String>(senderProps);
+		KafkaTemplate<Integer, String> template = new KafkaTemplate<>(pf);
+		template.setDefaultTopic(TEMPLATE_TOPIC);
+		template.syncConvertAndSend("foo");
+		assertThat(records.poll(10, TimeUnit.SECONDS), hasValue("foo"));
+		template.syncConvertAndSend(0, 2, "bar");
+		ConsumerRecord<Integer, String> received = records.poll(10, TimeUnit.SECONDS);
+		assertThat(received, hasKey(2));
+		assertThat(received, hasPartition(0));
+		assertThat(received, hasValue("bar"));
+		template.syncConvertAndSend(TEMPLATE_TOPIC, 0, 2, "baz");
+		received = records.poll(10, TimeUnit.SECONDS);
+		assertThat(received, hasKey(2));
+		assertThat(received, hasPartition(0));
+		assertThat(received, hasValue("baz"));
+	}
+
+}
+----
